HADOOP-2606. ReplicationMonitor selects data-nodes to replicate directly from needed replication blocks instead of looking up for the blocks for each live data-node. Contributed by Konstantin Shvachko.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/core/trunk@638655 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/dfs/PendingReplicationBlocks.java
@@ -100,7 +100,7 @@ class PendingReplicationBlocks {
   /**
    * The total number of blocks that are undergoing replication
    */
-  long size() {
+  int size() {
     return pendingReplications.size();
   }
